+++ server/http_error_handler_test.go
@@ -206,13 +206,12 @@ func TestErrorHandler_Backend(t *testing.T) {
 	}
 }
 
-func Test_StoreInvalidBackendResponse(t *testing.T) {
+func Test_ErroneousBackendResponse(t *testing.T) {
 	client := test.NewHTTPClient()
 
-	shutdown, hook := newCouper("testdata/integration/error_handler/06_couper.hcl", test.New(t))
-	defer shutdown()
-
 	type testcase struct {
+		name             string
+		file             string
 		path             string
 		expBody          string
 		expStatus        int
@@ -221,9 +220,13 @@ func Test_StoreInvalidBackendResponse(t *testing.T) {
 	}
 
 	for _, tc := range []testcase{
-		{"/anything", `{"req_path":"/anything","resp_ct":"application/json","resp_json_body_query":{},"resp_status":200}`, 418, 200, "status is not supported"},
+		{"store invalid backend response", "06_couper.hcl", "/anything", `{"req_path":"/anything","resp_ct":"application/json","resp_json_body_query":{},"resp_status":200}`, 418, 200, "status is not supported"},
+		{"api-level error handlers affect endpoint's buffer options", "08_couper.hcl", "/anything", `{"resp_json_status":200}`, 418, 200, ""},
 	} {
 		t.Run(tc.path, func(st *testing.T) {
+			shutdown, hook := newCouper("testdata/integration/error_handler/"+tc.file, test.New(t))
+			defer shutdown()
+
 			helper := test.New(st)
 			hook.Reset()
 
@@ -259,7 +262,11 @@ func Test_StoreInvalidBackendResponse(t *testing.T) {
 func getBackendLogStatusAndValidation(hook *logrustest.Hook) (int, string) {
 	for _, entry := range hook.AllEntries() {
 		if entry.Data["type"] == "couper_backend" {
-			return entry.Data["status"].(int), entry.Data["validation"].([]string)[0]
+			var validation string
+			if val, ok := entry.Data["validation"].([]string); ok && len(val) == 1 {
+				validation = val[0]
+			}
+			return entry.Data["status"].(int), validation
 		}
 	}